🟢 src/engines/prefilter.py -> Add PHashEngine for perceptual-hash based icon prefiltering, including folder filtering and dynamic thresholding. 🛠️ src/engines/ssim.py -> Partially refactored SSIMEngine to delegate prefilter, quality, and matching to new modular engines (PHashEngine, SSIMQualityEngine, SSIMMatchEngine). 🛠️ src/engines/ssim_common.py -> Extract common SSIM utilities (dynamic_hamming_cutoff, identify_overlay, multi_scale_match) into a shared module. 🟢 src/engines/ssim_match.py -> Partially refactored SSIMMatchEngine to handle parallel icon matching with main and fallback passes. 🟢 src/engines/ssim_quality.py -> Partially refactored SSIMQualityEngine for parallel overlay-based quality prediction per icon slot. 🛠️ src/iconmatch.py -> Update imports to register PHashEngine under "phash", clean up legacy engine comments, and adjust available options. 🛠️ src/sister.py -> Integrate new icon-matching stages into the pipeline: extend PipelineContext, update stage return types, and add quality detection, prefilter, and matching stages.
